注释

+++ src/instruction/instruction.go
@@ -2,21 +2,21 @@ package instruction
 
 import "fmt"
 
-// LuaInstruction Lua指令
+// LuaInstruction Lua指令（为一个32位无符号数）
 type LuaInstruction uint32
 
 // ABC 获取ABC模式下的参数（6，8，9，9）
 func (me LuaInstruction) ABC() (a, b, c int) {
-	a = int(me >> 6 & 0xff)
-	c = int(me >> 14 & 0x1ff)
-	b = int(me >> 23)
+	a = int(uint32(me) >> 6 & 0xff)
+	c = int(uint32(me) >> 14 & 0x1ff)
+	b = int(uint32(me) >> 23)
 	return a, b, c
 }
 
 // ABx 获取ABx模式下的参数（6，8，18）
 func (me LuaInstruction) ABx() (a, bx int) {
-	a = int(me >> 6 & 0xff)
-	bx = int(me >> 14)
+	a = int(uint32(me) >> 6 & 0xff)
+	bx = int(uint32(me) >> 14)
 	return a, bx
 }
 
@@ -35,7 +35,7 @@ func (me LuaInstruction) AsBx() (a, sbx int) {
 
 // Ax 获取Ax模式下的参数（6，26）
 func (me LuaInstruction) Ax() int {
-	return int(me >> 6)
+	return int(uint32(me) >> 6)
 }
 
 // OpCode 获取指令操作码（左边6个字节）
